[ServiceInfo] Corrigate frequency DVB-T/ATSC and add KSymbol/s
The Frequency of DVB-T and ATSC needs to be devided by 1000. Also ensure
we always have 3 digits after the decimal point. In addition add the
dimension for the bitrate (in KSymbol/s).

Depending on modulation type and FEC we also could consider to add bruto
and netto KB/s later

Thanks Huevos and Jpuigs for reporting the mistakes

http://forums.openpli.org/topic/44350-wrong-units-in-serviceinfo-using-dvb-t/